% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islets-package.R
\docType{package}
\name{islets-package}
\alias{islets}
\alias{islets-package}
\title{islets: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CamilleCoux/islets}
  \item Report bugs at \url{https://github.com/CamilleCoux/islets/issues}
}

}
\author{
\strong{Maintainer}: Camille Coux \email{camille.coux@orange.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
